requirements.txt
azure-cognitiveservices-speech~=1.36.0
boto3~=1.34.74
nltk~=3.8.1
PyAudio~=0.2.14
celery[redis]~=5.4.0
faster-whisper~=1.0.2
llama-cpp-python~=0.2.70
